Split tests into separate methods instead of using for loops
